Avoid integer overflow in av1_iidentity4/16/64_c()
BUG=b/74521979

+++ av1/common/av1_inv_txfm1d.c
@@ -1522,7 +1522,7 @@ void av1_iidentity4_c(const int32_t *input, int32_t *output, int8_t cos_bit,
   (void)cos_bit;
   (void)stage_range;
   for (int i = 0; i < 4; ++i)
-    output[i] = round_shift(NewSqrt2 * input[i], NewSqrt2Bits);
+    output[i] = round_shift((int64_t)NewSqrt2 * input[i], NewSqrt2Bits);
   assert(stage_range[0] + NewSqrt2Bits <= 32);
 }
 
@@ -1538,7 +1538,7 @@ void av1_iidentity16_c(const int32_t *input, int32_t *output, int8_t cos_bit,
   (void)cos_bit;
   (void)stage_range;
   for (int i = 0; i < 16; ++i)
-    output[i] = round_shift(NewSqrt2 * 2 * input[i], NewSqrt2Bits);
+    output[i] = round_shift((int64_t)NewSqrt2 * 2 * input[i], NewSqrt2Bits);
   assert(stage_range[0] + NewSqrt2Bits <= 32);
 }
 
@@ -1554,7 +1554,7 @@ void av1_iidentity64_c(const int32_t *input, int32_t *output, int8_t cos_bit,
   (void)cos_bit;
   (void)stage_range;
   for (int i = 0; i < 64; ++i)
-    output[i] = round_shift(NewSqrt2 * 4 * input[i], NewSqrt2Bits);
+    output[i] = round_shift((int64_t)NewSqrt2 * 4 * input[i], NewSqrt2Bits);
   assert(stage_range[0] + NewSqrt2Bits <= 32);
 }
